adjust text header
