+++ globoplay-quiz/HeaderView.swift
@@ -1,0 +1,36 @@
+//
+//  HeaderView.swift
+//  globoplay-quiz
+//
+//  Created by eduardo.silva on 07/11/19.
+//  Copyright © 2019 Globo. All rights reserved.
+//
+
+import UIKit
+
+final class HeaderView: UICollectionReusableView {
+    
+    static let identifier: String = "Header"
+    private let label: UILabel = {
+        let l = UILabel()
+        l.translatesAutoresizingMaskIntoConstraints = false
+        return l
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        self.addSubview(label)
+        label.topAnchor.constraint(equalTo: self.topAnchor, constant: 8).isActive = true
+        label.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
+        label.trailingAnchor.constraint(equalTo: self.trailingAnchor).isActive = true
+        label.leadingAnchor.constraint(equalTo: self.leadingAnchor).isActive = true
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func setup(_ desc: String) {
+        label.text = desc
+    }
+}
